feat(email connector): put everything inside a  directory
