+++ connectors/email/src/main/java/io/camunda/connector/email/outbound/protocols/config/CryptographicProtocol.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. Licensed under a proprietary license.
+ * See the License.txt file for more information. You may not use this file
+ * except in compliance with the proprietary license.
+ */
+package io.camunda.connector.email.outbound.protocols.config;
+
+public enum CryptographicProtocol {
+  NONE,
+  TLS,
+  SSL;
+}
