--- connectors/email/src/main/java/io/camunda/connector/email/protocols/actions/Action.java
@@ -1,9 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. Licensed under a proprietary license.
- * See the License.txt file for more information. You may not use this file
- * except in compliance with the proprietary license.
- */
-package io.camunda.connector.email.protocols.actions;
-
-public sealed interface Action permits SmtpAction, Pop3Action, ImapAction {}
