--- connectors/email/src/main/java/io/camunda/connector/email/response/Pop3ListEmailsResponse.java
@@ -1,12 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. Licensed under a proprietary license.
- * See the License.txt file for more information. You may not use this file
- * except in compliance with the proprietary license.
- */
-package io.camunda.connector.email.response;
-
-import java.util.List;
-
-public record Pop3ListEmailsResponse(
-    String uidl, List<String> fromAddresses, String subject, Integer size) {}
