+++ connectors/email/src/main/resources/META-INF/services/io.camunda.connector.api.outbound.OutboundConnectorFunction
@@ -1,1 +1,1 @@
-io.camunda.connector.email.EmailConnectorFunction
+io.camunda.connector.email.outbound.EmailConnectorFunction
