--- connectors/email/src/test/java/io/camunda/connector/email/core/JakartaExecutorTest.java
@@ -1,15 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. Licensed under a proprietary license.
- * See the License.txt file for more information. You may not use this file
- * except in compliance with the proprietary license.
- */
-package io.camunda.connector.email.core;
-
-import org.junit.jupiter.api.Test;
-
-class JakartaExecutorTest {
-
-  @Test
-  void execute() {}
-}
